Regroup file and path helpers
