+++ packages/samen/src/production-server.ts
@@ -2,7 +2,7 @@
 
 import { Environment, handleError } from "@samen/core"
 import serve from "./commands/serve"
-import getManifestPath from "./getManifestPath"
+import { getManifestPath } from "./utils/paths"
 
 process.on("unhandledRejection", handleError)
